requirements.txt
fastapi==0.115.12
uvicorn==0.22.0
openai==0.27.0
psycopg2-binary==2.9.3
pydantic-settings==2.8.1;
python-jose[cryptography]==3.4.0 
passlib==1.7.4
celery==5.4.0
redis==5.2.1
alembic==1.15.1
python-multipart==0.0.20
bcrypt==3.2.2
eventlet==0.39.1
cryptography==39.0.1
pypdf==5.4.0
pytest==7.1.2 
pytest-asyncio==0.18.3 
httpx==0.28.1
pytest-cov==6.0.0
asyncpg==0.30.0